+++ Rdmp.UI/Menus/CohortIdentificationConfigurationMenu.cs
@@ -53,7 +53,7 @@ namespace Rdmp.UI.Menus
             Add(new ExecuteCommandCreateNewCohortIdentificationConfiguration(_activator));
 
             if(_activator.CoreChildProvider is DataExportChildProvider dx)
-                AddGoTo(dx.AllProjectAssociatedCics.Where(a=>a.CohortIdentificationConfiguration_ID == cic.ID).Select(a=>a.Project).Distinct(),"Project(s)");
+                AddGoTo(()=>dx.AllProjectAssociatedCics.Where(a=>a.CohortIdentificationConfiguration_ID == cic.ID).Select(a=>a.Project).Distinct(),"Project(s)");
             
         }