+++ CHANGELOG.md
@@ -4,11 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+...
+
+## [3.1.0]
 
 ### Added
 
 - Cohort sets with HAVING sql now support 'View Dataset Sample' (of matched records)
+- Added new property IsView to TableInfo
+- Added GoTo menu item Catalogue=>TableInfo
 - Added user setting for skipping Cohort Creation wizard
 - MDFAttacher emits more messages when looking up location on disk to copy MDF file to.
 - Added menu option to set IsExtractionIdentifier on a Catalogue without having to open ExtractionInformations directly
@@ -35,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where deleting a Filter from a cohort set in a Cohort Identification Query could result in the display order changing to alphabetical (until tab was refreshed).
 - Fixed obscure bug in plugins implementing the `ICustomUI` interface when returning a new object in `GetFinalStateOfUnderlyingObject` that resulted in the UI showing a stale version of the object
 - Connecting to a non existant server in ServerDatabaseTableSelector now shows the Exception in the RAG icon (previously just showed empty database list)
+ 
 - Fixed bug where adding/removing a column in Aggregate Editor would would reset the Name/Description if there were unsaved changes (to Name/Description)
 - Fixed bug where example datasets created would have the text value "NULL" instead of db nulls (only affected initial install/setup datasets)
 
@@ -109,7 +117,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Culture (e.g. en-us) not being passed correctly in DelimitedFlatFileAttacher
 - Fixed bug where Updater would show older versions of RDMP as installable 'updates'
 
-[Unreleased]: https://github.com/HicServices/RDMP/compare/v3.0.16-rc2...develop
+[Unreleased]: https://github.com/HicServices/RDMP/compare/release/3.1.0...develop
+[3.1.0]: https://github.com/HicServices/RDMP/compare/v3.0.16-rc2...release/3.1.0
 [3.0.16-rc2]: https://github.com/HicServices/RDMP/compare/v3.0.16-rc...v3.0.16-rc2
 [3.0.16-rc]: https://github.com/HicServices/RDMP/compare/v3.0.15...v3.0.16-rc
 [FAnsiSql]: https://github.com/HicServices/FAnsiSql/